WIP First demo version of upcoming xcp-metrics initial PR
Initial PR contains only the squeezed plugin, to replace
xcp-rrdp-squeezed.

This is a non-standard packaging, explicitly not following the
Fedora/RH guidelines, so we can support a large number of RPM
distro versions without having to deal individually with the 200+
dependency crates.  The source package includes all of them in
one go.

Sources for the package are the source tarball itself, and a
tarball of all dependencies obtained by "cargo vendor".  In this
version it is complicated by the fact we're using an unreleased
version of xenstore-rs, with changes, and "cargo vendor" does not
support it, so:
- the vendor tarball includes the official xenstore-rs-0.3.2
- our additional features in xenstore-rs come in a patch
- since "cargo release" messes with Cargo.toml (and we don't need
  this mess) we first apply a patch reverting to the 0.3.2 git version
  of xenstore-rs Cargo.toml, so our feature patch can apply

FIXME: need to get git tree for patch crate in vendor tree